@@ -1,5 +1,4 @@
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace ApiTaCerto.Models.Usuario
 {
@@ -11,19 +10,13 @@ namespace ApiTaCerto.Models.Usuario
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.HasDefaultSchema("TaCerto");
-            foreach (var entityType in modelBuilder.Model.GetEntityTypes())
-            {
-                // equivalent of modelBuilder.Conventions.Remove<PluralizingTableNameConvention>();
-                entityType.Relational().TableName = entityType.DisplayName();
-            }
         }
 
         public DbSet<Pessoa> Pessoas { get; set; }
         public DbSet<PessoaToken> PessoaToken { get; set; }
         public DbSet<Disciplina> Disciplinas { get; set; }
         public DbSet<Instituicao> Instituicoes { get; set; }
-        public DbQuery<AtividadeDisciplina> AtividadeDisciplinas { get; set; }
+        //public DbSet<AtividadeDisciplina> AtividadeDisciplinas { get; set; }
         public DbSet<Atividade> Atividades { get; set; }
         public DbSet<Questao> Questoes { get; set; }
         public DbSet<TurmaAluno> TurmaAlunos { get; set; }
