@@ -37,7 +37,8 @@ namespace ApiTaCerto.Repositorio
         }
 
         public IEnumerable<AtividadeDisciplina> GetAllClassActivities(long idDisciplina){ 
-            return _contexto.AtividadeDisciplinas.Where(ad => ad.IdDisciplina == idDisciplina).ToList();
+            //return _contexto.AtividadeDisciplinas.Where(ad => ad.IdDisciplina == idDisciplina).ToList();
+            return null;
         }
 
         public IEnumerable<Atividade> GetAllActivities(long idDisciplinaAutor){
